ai-core/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.55.0
python-multipart>=0.0.12
aiofiles>=24.1.0
python-dotenv>=1.0.0
sqlite-vec>=0.1.0
pypdf>=5.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=1.0.0
markitdown>=0.0.1
json5>=0.9.0
faster-whisper>=1.0.0
pydub>=0.25.0
imageio-ffmpeg>=0.4.0
moviepy>=1.0.3
httpx>=0.28.0
sse-starlette>=2.1.0
paddlepaddle==3.0.0
paddleocr>=2.9.0
pypdfium2>=4.30.0
Pillow>=10.0.0